Extract understanding of "better" and "worse" to the Status type.
This uses the suggestion from @wizzwizz4@fosstodon.org to derive Ord.

The point here is so that the 'solve' code can conceptually represent
the high level logic of "try all the moves and find the best one". The
idea of "best" is represented by deriving Ord for Status (and putting
the enum values in order from worst to best). The 'complement' method is
defined to make the logic in solve() a little shorter - it encodes the
idea that a Win for our opponent is a Loss for us and vice-versa.

Even this makes it faster?? Seriously I don't know what's going on here.

solve (new Board)       time:   [826.36 µs 827.69 µs 829.17 µs]
                        change: [-43.738% -43.325% -42.907%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 10 outliers among 100 measurements (10.00%)
  2 (2.00%) low mild
  2 (2.00%) high mild
  6 (6.00%) high severe