Remove InvalidMove error path
Since the bit twiddling code only tries moves that are 'valid', we no
longer need the error path and the check for it. It is replaced with
some debug assertions so that in debug builds we have a chance of
catching problems if solve() tries an invalid move (which would strictly
be a bug).

solve (new Board)       time:   [556.94 µs 557.93 µs 558.98 µs]
                        change: [-62.128% -61.859% -61.601%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 500 measurements (0.60%)
  3 (0.60%) high mild